# Conflicts:
#	pom.xml
#	src/main/java/com/example/demo/auth/controller/AuthController.java
#	src/main/java/com/example/demo/auth/controller/UsuarioController.java
#	src/main/java/com/example/demo/auth/dto/RegisterRequest.java
#	src/main/java/com/example/demo/auth/dto/UsuarioDTO.java
#	src/main/java/com/example/demo/auth/entity/Usuario.java
#	src/main/java/com/example/demo/auth/jwt/JwtAuthenticationFilter.java
#	src/main/java/com/example/demo/auth/jwt/JwtService.java
#	src/main/java/com/example/demo/auth/repository/UsuarioRepository.java
#	src/main/java/com/example/demo/auth/servicios/UsuarioService.java
#	src/main/java/com/example/demo/config/SecurityConfig.java
#	src/main/java/com/example/demo/controller/Costo_ServicioController.java
#	src/main/java/com/example/demo/dto/EmpleadoDTO.java
#	src/main/java/com/example/demo/dto/crear/UsuarioCrearDTO.java
#	src/main/java/com/example/demo/mapper/noIdenticos/EmpleadoCrearMapper.java
#	src/main/java/com/example/demo/mapper/noIdenticos/PasajeroCrearMapper.java
#	src/main/java/com/example/demo/mapper/noIdenticos/UsuarioCrearMapper.java
#	src/main/java/com/example/demo/model/Credenciales.java
#	src/main/resources/application.properties

+++ src/main/java/com/example/demo/dto/CostoServicioDTO.java
@@ -21,6 +21,6 @@ public class CostoServicioDTO {
 
     @DecimalMin(value = "0.0", message = "El costo no puede ser negativo")
     @Digits(integer = 6, fraction = 2, message = "El costo debe tener máximo 6 dígitos enteros y 2 decimales")
-    private BigDecimal costo;
+    private Double costo;
 
 }
